+++ packages/@leight/query/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leight/query
 
+## 0.6.2
+
+### Patch Changes
+
+- Improved React Query stuff
+
 ## 0.6.1
 
 ### Patch Changes